requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
pdf2image
langchain-core
langchain-google-genai
poppler-utils
elevenlabs
python-multipart